move TickResults to Cardinal
